+++ api_steward-rails/lib/api_steward/rails.rb
@@ -3,6 +3,8 @@
 require "api_steward"
 require "api_steward/rails/version"
 require "api_steward/rails/railtie"
+require "api_steward/rails/identify"
+require "api_steward/rails/govern"
 
 module ApiSteward
   # Rails glue for api_steward. Bundler requires this automatically for the